feat: add typeorm, pg and implement User

+++ src/user/user.controller.ts
@@ -8,6 +8,7 @@ import {
   Delete,
   HttpCode,
   NotFoundException,
+  ForbiddenException,
   ParseUUIDPipe,
 } from '@nestjs/common';
 import {
@@ -28,6 +29,12 @@ import { CreateUserDto } from './dto/create-user.dto';
 import { UpdatePasswordDto } from './dto/update-user.dto';
 import { User } from './entities/user.entity';
 
+const stripPassword = (user: User) => {
+  const clone = Object.assign({}, user);
+  delete clone.password;
+  return clone;
+};
+
 @Controller('user')
 @ApiTags('User')
 export class UserController {
@@ -41,8 +48,9 @@ export class UserController {
     description: 'Users successfully getted',
     type: [User],
   })
-  findAll() {
-    return this.userService.findAll();
+  async findAll() {
+    const allUsers = await this.userService.findAll();
+    return allUsers.map((user) => stripPassword(user));
   }
 
   @Get(':id')
@@ -65,12 +73,12 @@ export class UserController {
   @ApiNotFoundResponse({
     description: 'User with given id not found',
   })
-  findOne(@Param('id', new ParseUUIDPipe({ version: '4' })) id: string) {
-    const user = this.userService.findOne(id);
+  async findOne(@Param('id', new ParseUUIDPipe({ version: '4' })) id: string) {
+    const user = await this.userService.findOne(id);
     if (!user) {
       throw new NotFoundException(`User with id ${id} not found`);
     }
-    return user;
+    return stripPassword(user);
   }
 
   @Post()
@@ -88,8 +96,8 @@ export class UserController {
   @ApiBadRequestResponse({
     description: 'Request body does not contain required fields',
   })
-  create(@Body() createUserDto: CreateUserDto) {
-    return this.userService.create(createUserDto);
+  async create(@Body() createUserDto: CreateUserDto) {
+    return stripPassword(await this.userService.create(createUserDto));
   }
 
   @Put(':id')
@@ -115,15 +123,23 @@ export class UserController {
   })
   @ApiNotFoundResponse({ description: 'User with given id not found' })
   @ApiForbiddenResponse({ description: 'Old password is wrong' })
-  update(
+  async update(
     @Param('id', new ParseUUIDPipe({ version: '4' })) id: string,
     @Body() updateUserDto: UpdatePasswordDto,
   ) {
-    const user = this.userService.update(id, updateUserDto);
+    const user = await this.userService.findOne(id);
     if (!user) {
-      throw new NotFoundException(`User with id ${id} not found`);
+      throw new NotFoundException({
+        message: `User with id ${id} is not found`,
+      });
     }
-    return user;
+    if (user.password !== updateUserDto.oldPassword) {
+      throw new ForbiddenException(
+        `Old password is wrong for user with id ${id}`,
+      );
+    }
+    const userUpd = await this.userService.update(id, updateUserDto);
+    return stripPassword(userUpd);
   }
 
   @Delete(':id')
@@ -144,7 +160,13 @@ export class UserController {
   })
   @ApiNotFoundResponse({ description: 'User with given id not found' })
   @HttpCode(204)
-  remove(@Param('id', new ParseUUIDPipe({ version: '4' })) id: string) {
-    return this.userService.remove(id);
+  async remove(@Param('id', new ParseUUIDPipe({ version: '4' })) id: string) {
+    const user = await this.userService.findOne(id);
+    if (!user) {
+      throw new NotFoundException({
+        message: `User with id ${id} is not found`,
+      });
+    }
+    return await this.userService.remove(id);
   }
 }
